RESULTS : Model overfit
BODY :

The training loss and accuracy approached 0 and 100, respectively.  However
the descent/ascent did not begin until the 3rd epoch.  Validation loss
ascended, indicating overfitting; again, this ascent did not start until the
3rd epoch.  Validation accuracy stuck at ~40%, and it appears that
validation set predictions were not changing from test to test for the final
epoch, similar to the pattern seen throughout the last run.  There was very
little learning for the first 3 epochs, with the loss and accuracy for both
training and validation sets remaining constant.

I did use the AdaGrad optimizer for this run, and did not retrieve the learning
rate or other parameters.  It is possible a shift in these numbers caused the
shift from learning to quick overfit.

An early version of this code, in which the activation functions between the
convolutional layers was linear, not Relu, appeared to learn gradually, with
loss and accuracy shifting at the same rates for both the training and test set.

The next change to the model I will test is the activation function for the
convolutional layers.  Also I will attempt to pull the relevant statistics
from the AdaGrad optimizer for greater insight into the results of the run.