RESULT : Abject failure - accuracy never improves
BODY :

There were many issues with this model.  First, the validation prediction
appeared to get stuck into the same patterns after about 11000 images observed.
This is similar to the pattern seen in the previous model version.  Does this
represent the model reaching a local minima for training loss?  Possibly, but
the loss for both the training and the validation set continued to decrease
through two epochs, at which point the run was aborted.  To my mind, if the
loss is decreasing, there should be variation in validation predictions.  Perhaps
under the hood, the model is predicting the same class repeatedly.  It may be
helpful to add print statements to the session in order to observe the exact
predictions of the validation set on occasion.

Another issue with this run is the massive difference between training loss
(which plateaus at about 1e+9), and the validation loss which quickly dropped
to 1e+5, and continued to decrease gradually from there.  This does not make
sense.  How can the loss derived from unseen, untrained images be so much
less than the training image set, which should eventually be learned verbatim.
I feel there may be an error in the labeling or separation of the training
and validation sets.

Probable amendments for the next model :
     * reduce beta parameter in regularization to provide it influence, but not
          dominance over the loss term.
     * reduce the number of layers in the model - it may simply be too complex
     * reduce the pixel size of the images until learning improves to something moderate.
     * refactor the split of the training and validation sets.  Separate these image filenames
          prior to balancing via duplication, then mutation.
     * refactor the standardization function.  Would -1 to 1 be more appropriate than -0.5 to 0.5?
     * (possible) add dialation to convolution steps
     * change kernel for every convolution to 3x3, and each pooling layer keeps the
          same depth until the first convolution after pooling.

In conclusion, the model is currently not functional.  It needs a major revision
and likely a significant simplification before it will predict fish classes with
any accuracy.