docs : Model does not generalize
BODY :

In this model run, the validation set was made up from the unbalanced original
data set.  Images were uploaded into array format, standardized, but not
mutated.  The training set was balanced by duplicating filenames as needed
to give each fish class an equivalent number of images.  From there images
were converted to array, standardized, and mutated randomly for each training
instance.  Thus no exact image should have been seen more than once in the
training set, and the images in the validation set would be representative of
the test training set to be seen later.

Because all of the training images are mutated to prevent model variance, yet
the validation set was left in original state, I predicted that the loss and
accuracy measurements may exhibit successful learning prior to similar exhibitions
on the training set.  Indeed, this seemed to be the case for the first 2 epochs.
The accuracy of training set prediction remained flat at about 17-18%, whereas
the validation set accuracy increased to 22%.  The loss statistics for both the
training and test sets exhibited a 2-log decrease (10000 to 200) in the first 1000 images, prior
to very gradual decrease (200 to 160 over 74000 images).  This indicates the
learning rate may have been set too high for this run.  Nevertheless, training
was still occuring gradually after 5 epochs.

The odd thing about this run is that after 2 epochs, the training set accuracy
began to increase and flattened at 27%.  At the same time, the validation set
accuracty decreased to 7% and exhibited the same pattern of prediction in
every single validation measurement from the start of the 3rd epoch.  Again,
it seems the model gets stuck at a sub-optimal local minima.

Parameters to change:
     * Increase validation set to 250
     * Decrease the learning rate from 1e-2 to 1e-4.
     * Change the depth of convolution layers to thinner at the end of convolution.

These changes should help assess whether overfitting is occuring and the reason
for these inconsistent learning results.