RESULT : v1.3.0
BODY :

The new filename input system is functional, and a bit faster than previous
versions of the code.  Importantly, the validation set can be run in a series
of batches to return a single cross_entropy or accuracy score, resulting in
more consistent and helpful data trends.  Previous code versions consumed
75,000 images overnight and exhibited signs of overfitting.  This model version
consumed ~115,000 images overnight and both the training and validation sets
were both decreasing in loss and increasing in accuracy when the number of
epochs was eclipsed, albeit gradually in both metrics.

This is a great benchmark run, as no mutation or dataset balancing was implemented.
The new input pipeline appears much faster and will allow a more clear distinction between
different models.

Some ideas for next modifications:
     * dilation in first convolution
     * print filters?
     * ReLU -> ELU or another activation fn that maintains negative activations
     * add mutation to training set
     * Implement test image predictions and upload to Kaggle!