RESULT v1.3.3.3 :
BODY :

Increasing the learning rate caused the validation set accuracy to get stuck
at 0.4350, the exact percentage it got stuck on in v1.3.3(.0).  I am unsure
what is going on.

I tried to submit the test logits to Kaggle, but the Kaggle upload system does
not normalize in order to avoid large numbers and an error was tripped.  I
will compute the softmax for the logits from now on.

That said, when I computed the softmax for this set of test logits, the results
were very aggressive.  1.0 versus 0.0's for all other classes in almost every
case.

Conclusions :

There may be a bug in the way I am detecting accuracy.  The aggressive nature
of the model could indicate why there is no variation in the validation set
predictions, as incremental changes in the model weights are unable to move
predictions out of extreme classification paths.  Perhaps using relu6 would
help prevent this runaway.