RESULT v1.3.3.5 : Same weird consistent accuracy in the validation set
BODY :

Decreasing the decay rate to 0.1 instead of 0.9 did not have any effect on the
outcome of training.  The validation set accuracy gets stuck very early in
the training process. My guess is that the learning rate is too high, but in
theory, an elevated learning rate should eventually cause the validation set
accuracy to decrease, I think.  Or it is possible that under the hood, the
network is predicting the same class every instance.

However in the test set softmax logits, at least some of the images were classified
with distributed probabilities, unlike the last run in which all test images
were aggressively predicted as 0 or 1.  For the heck of it, I will submit this
to Kaggle to get a reference point, though I expect a very poor performance measure