refactor : Moved parameters to separate file in repo
BODY :

Parameters moved to a separate file, with the MODEL.ipynb now serving only as
a wrapper for the other python scripts.  This was done to make the tracking of
parameter changes easier.  Right now updated versions of MODEL could be due to
parameter changes or to having run a graph session.