RESULT v2.0.0.0 : Local minima?
BODY :

Validation set accuracy stuck at 42% from step 3 onwards.  Validation loss
converged to 1.600 and no further.  Validation logits are proportional to the
fish frequencies.

Analysis of the test_df revealed that every record also exhibited probabilities
that were proportional to fish frequency.  Weirdly, the first 25 entries are
nearly identical, the second 25 entries exhibit this pattern but are shifted
slightly from the first 25.  This pattern continues throughout.  'ALB' is always
the most probable prediction, but its probability shifts from batch of 25 to batch
of 25.  This should not be.  There may be a bug in the queue system or it does
not work as I thought it did.  If there is a bug in this set-up, there may be
a bug in the validation set-up as well.

Finally, the distributions of the W_variables did not seem to be updated, except
for slightly in W_conv5.  Perhaps the learning rate was too low.

Next steps :
     * v2.0.0.1 will be a brief run with a very high and very rapidly decreasing
     learing rate in order to test the local minima theory.

     * Debug test 'batch effects' later