feat : Process images on CPU ; Bias terms initialized at zero
BODY :

Queue input reading system replaced with fd.process_batch prior to the session
call.  This was done in a debugging effort, and is functional, however the
cause of the bug was unrelated to the queue input reading system.  I need to
experiment for which file read system is faster or more elegant.  For now this
set-up is adequate.

The cause of the bug whereby different images produced the exact same logits
was due to the initialization of convolution bias terms to 1, instead of zero.
Initializing at 1 was described in the ImageNet publication, however in my
model, the ones were dominating any activations.  Thus no matter the composition
of an image, the bias terms were controlling the output of the model.

Model learns as expected once these bias terms were set to zero at the init call.