RESULT v2.1.0.0 : tensorboard --logdir='TB_logs/v2.1.0.0'
BODY :

Project code completed without error and generated a csv file for upload at
the Kaggle portal.  12 epochs of the balanced data set (~160,000 images) were
consumed in just over 18 hours.  Both validation accuracy and cross_entropy
were still dynamic after the 12 epoch, suggesting room for learning is available
if the training were run longer.

Compared to v2.0.0.5, which was the first functional version of ImageNet architecture,
v2.1.0.0 learned at a slower rate. There were more inactivated ('dead') nodes
after the first convolution in the new version.  Dead nodes were not tracked
after the first convolution in v2.0.0.5, however ~75% of nodes were inactive after
conv2 through conv4.  And convolution layer 5 was 98-99% inactive.  It is possible
that the saturation of dead nodes contributed to the slow learning observed.  A
difference in the two versions that could have caused this is the normalization
strategy.  I used a pixel offset and pixel factor of 100, 100 in this model, resulting
in a validation mean pixel value of -0.01 +/- 0.53.  A slightly different strategy
was taken for v2.0.0.5, resulting in mean pixel values of ~0.001 +/- 0.35.  I
hypothesize that this slight discrepancy in mean pixel value caused fewer
relu activated nodes in the fist convolution, which had a knock-on effect, such
that very little information was emitted from the final convolution layer.  I would
have expected the bias terms to eventually rectify this, and indeed the number
of activated nodes was increasing for all convolutional layers at the time of
epoch completion.  So it is possible that with a better normalization strategy
and longer training duration, this issue can be solved.

Next steps :
     * Normalization such that the mean pixel value is slightly positive
     * 20 epochs
     * Test preds are mean of 5 crops
     * Experiment with slightly higher learning rate.
     * Implement a time / image statistic ??

Conclusion :
This version is the most accurate model yet, and, excitingly, there appears to be
room for improvement.