RESULT v2.2.1.0
BODY :

The main change for this model was the use of elu for the first and fifth
convolutional layers.  This was done to maintain information throughout the
process, as the final convolutional layer in v2.1.0.0 was activated in only 1-2%
of nodes.  In this run, convolutional layers 3 and 4 were now only activated
in <10% of nodes.  However the model did train slowly and validation
accuracy was still increasing after 25 epochs had been consumed.