RESULT v2.2.1.2 : Best validation accuracy but test set does not achieive usable score
BODY :

After training this particular model (v2 series) for about 54 hours and
over 600,000 images, a validation accuracy of 80%, and a cross_entropy
score of 0.6 were achieved.  These are best scores thus far.  Considering
the training images were balanced and the validation images were native
(significantly unbalanced), I think this represents good learning of the
training set.

Unfortunately the model did not generalize well to the test set stage
1 images.  The metric for loss increased proportionally with training and
the final model achieved a log loss score of >2.0.  By simply using the
frequencies of class in the training set, a benchmark of 1.6 can be
achieved so this model is worse than a systematic guess.

Why is this?  Cursory analysis of the test set stage 1 images reveals
images from new boats or on different days.  Thus it appears my training set
with removed validation images are significantly different from the test
set images.  It is likely my model is learning feature representations
such as the boat or the time of day (overall brightness) instead of the
fish.  Also, the 'other' was not learned as well as other classes for
some reason.

With this in mind I will move on to a new strategy for version 3 series,
in which an autoencoder will be trained and used to label individual
cropped fovea via a one-shot learning propigation of bona fide labels.
This should allow more control over what I would like the model to
look for, although will lose the ability to find hidden latent features
(such as the boat or fisherman) associated with each fish class (this is
likely to be a good thing and make the model generalize better).  It
will be a complicated project and set-up much differently to this
current network.

CONCLUSION :
The test set images are significantly different from the training set
images that a different strategy is required.  Colinearity in the training
set images between boats and fisherman and fish class are not maintained
in the test set images, and as such, the model fails to generalize.
In any case a successful model was built that learned UNSEEN images
as the validation images were separated from the training set PRIOR
to balancing and training.