RESULT v3.5.6.0 : Functioning Autoencoder
BODY :
After exploring different graph architectures, activation strategies,
layer depths, etc, a simple autoencoder was generated.  Images could
be 'compressed' into a latent layer and decoded into a recognizable
representation of the original image.  The decoded images were usually
pixelated near lines or changes in color, however swathes of colored
space were adequately represented.  The problem with this autoencoder
is that the encoded layer actually contains more nodes than there
were pixels in the original image.  Thus it does not actually
compress the data at all.  This is a proof of principle step to serve
as a foundation for the next steps, in which more layers will be
added with striding that will gradually reduce the number of nodes into
the final encoded layer.  Morever the encoding / decoding process provides
a qualitative check of what image structures are being preserved in the
'compression'.

NOTE : the most recent updates the graph were made out of turn to this
result commit.

ref : v3.5.6.0