RESULT v3.6.0.0 : Weak learning
BODY :
This simple model, in which depths were restricted to the input depth of
3, for every layer, failed to encode decode with great precision.
The outlines of structures of the original image can be seen weakly,
however little difference in each channel is observed and the final
output decoded image is always grey.  I will increase the depth of
internal layers slightly and re-run this first epoch.

ref : v3.6.0.0