RESULT v3.6.1.0 : Learning rate is too low
BODY :

Increasing the depth of internal layers had a slight effect on the rate
of learning.  However, major advances were made when the learning rate
was increased to 0.1.  After this change in the session calls,
output images quickly began to represent the input and the log loss
plummeted.  I will slightly increase the depth of internal layers AND
start the learning at 0.1 this version.

ref : v3.6.1.0