RESULT v3.6.3.0 : No improvement over previous version
BODY :

Two changes were made for this version:
     * relu activation for encoding convolutions.
     * Increase depth of 2nd encoding convolution from 16 to 32.

The first deconvolution was linearly activated, and the final deconvolution
was left with sigmoid activation.  The log_loss exhibited a similar pattern
to the previous version, but did not eclipse the minimum log_loss observed
with slightly smaller depth and linear activations.  It's possible this
slight decrease in performance is due to a loss of information, with negative
tensor values squashed during the encoding steps.

The output image representations are structurally similar, but do not
exhibit the same color patterns as the original.  I am unsure if this is
important.  It could be, with the LAG fish exhibiting red coloring, however
structure may be the most important representation going forward.

For one last trial, in the v3.6 series, I will increase the depths of
the convolution / deconvolution layers.  Hopefully this increase in
information retained will allow encoding of the color patterns.  After this
run, the best architecture will be used for a 3 layer encoding and
analogous decoding.

ref : v3.6.3.0