feat : Propigate fovea labels upon confident KNN prediction
BODY :

A function was generated to display the segments of high-resolution
images (called fovea) from coordinates generated in the FOVEA_EMBEDDING.ipynb
file.  This function prompted for labels for each fovea, 'F' for fish or
'N' for 'NoF'.  When 'F' was received, the fovea label was set to the
identity of the image label.  When 'N' was received, a fovea label of
'NoF' was defined.  During this process, in which random high-resolution
images from each label class were retrieved, graphical representation
of corresponding embeddings were visualized in a variety of different
dimension reduction techniques.  In total, it evident that some fish
classes were easily distinguished from background ('NoF') areas based
on their embedding vectors.  Other fish classes were more difficult to
distinguish in this manner.

Because of this discrepency, I was not able to propigate fish class
labels confidently to the whole fovea set.  However, I was able to
propigate the fovea within each high-resolution image that definitely
did not contain the fish.  To accomplish this, a KNN classifier was
generated and trained on the bona fide images from each specific
image-label class.  For example, all the images from the 'ALB' set were
retrieved, and then split into those that had been assessed and those where
fovea labels remained unknown.  The KNN was trained on the bona fide labels,
and then used to predict 'NoF' or 'ALB' in the 'Unknown'.  Probabilities that
the fovea class was 'NoF' that exceeded 0.94 were propigated to the fovea set
labels.

Conclusion :
This autoencoding plus one-shot labelling allowed a KNN classifier to
confidently identify fovea from each image class which were certainly
not the fish in question.  The resulting fovea set will be used as the
starting training set for a semi-supervised classification network that
will add training images as predictions grow confident.