feat : FISH_FINDER model is functional
BODY :

A pilot run overnight was successfully implemented.  Cross entropy
decreased and the number of staged fovea grew over the epochs.  Files
were successfully stored and retrieved as needed.  The stop.txt
clause worked as intended.

Regarding the model and parameter choices, the learning rate could
probably be increased slightly as very little movement in the W
weights was observed, while the bias terms were still migrating after
7 epochs.  That said, several starts of this model were unstable,
so it is possible a low learning rate is required for some reason.

Speed - the model trained on 93000 images overnight (10 hours).  The
model checkpoints will allow further training to be picked up at this
point.

Stage and commit to training set - no fovea were committed to the
training set yet.  So this must be tested.  Also, test for picking the
model back up in subsequent runs must be made.

Conclusion - this was a successful pilot run and the next step is
optimization of the stage -> training set commitment phase.

ref : PILOT_RUN_v1