Feat : Implement backdoor to training process
BODY :

This script allows the user to navigate through fovea that have been
saved in the stage dictionary, and commit them to the actual training
set if accurate.  Moreover, user can override the predicted label
to send labeled fovea without the model having to be correct.  Eventually
this process should be governed automatically, but predictions are not
yet stable or accurate enough to turn loose to an algorithm.

ref : PILOT_RUN_v5.0