Fix : Add Mode flag to every instance of misc.imresize
BODY :

I may have found a bug by where the misc.imresize function automatically
scales pixel values to uint8 scale from 0 to 255.  This process occurs
if a dtype is not pre-cast into the incoming array.  I think this is
taken care of by 'RGB' mode call in misc.imread function.  However,
if an image with floats is provided, or if the type is ambiguous, to
misc.imresize, the function will scale the minimum value to 0 and the
maximum value to 255.

In such case, an image without a minimum of 0 or maximum of 255 was not
present in the image,pixel values would be scaled incorrectly compared
to the entire image set.  I have now added the mode = 'RGB' tag to each
instance of misc.resize to indicate that image should not be rescaled.
The normalization process in `process_fovea`
scales convers the array to float anyway, in the process of scaling to
the float range of -0.5 to 0.5.