Feat : Add high-res set up and prediction phase
BODY :

New function in fish_data.py for the stacking of fovea from a single
high-resolution image.  Stack of fovea will be fed to FISHFINDER model
for prediction of each fovea.  This tensor of predictions will be
fetched in the session and subjected to further modelling for accurate
and precise classification of high resolution image.