Feat : Run pilot predictions of high-res images
BODY :

Having trained the FishFinder model extensively, a function was
written to parse a high-resolution image into a stack of fovea.  Fovea
were then each run through the FishFinder model and reassembled into
a prediction array where each y, x position roughly represents the
location of the original high-resolution image.  From this
cursory analysis, it is clear that the model is good at predicting
locations of 'NoF' and locations where a fish is present.  However it
fails to correctly identify the type of fish, with predictions mimicking
the ratio of foveal labels which were trained upon.  Thus the general
architecture works but the cost function to implore the model to learn
the less frequent types of fish did not.

Nevertheless this was a successful pilot experiment and will be used in
the next version of the model.  In general, the autoencoder section
will be skipped to a manually labeled fovea set.  The scale for resizing
was actually too low, and should be above 1 in most instances.  By
optimizing the scale size and the fovea_size, I think I can speed model
training up.  However, the most important change will be to use a
balanced  training set.

A note on version control, I am going to merge the pilot into the master
and then clear the repo, adding back fresh versions of each component.
This is more for my sanity than necessity.

ref : PILOT_RUN_v5