Feat : Working model generated
BODY : Fovea from the training set are input to train the model in
a balanced fashion.  Control over the learning rate, stop training,
is accomplished with extratenous txt files.  Upon prompt, the model
will iterate through each high-resolution image and produce a stack
of fovea predictions, which can be sent through a standard ML algorithm
for classification later. The model is training very slowly, so it is
possible further optimization is required.

Experimentation with allowing the convolution layers to train, or for
the fully connected network to be activated via a different function
besides relu.  Right now predictions for every image are simply the
proportion of images seen by the model. Changes in fovea have no affect
on the output logits.  Perhaps increasing the learning rate will help
as well.